LayoutTests/touchadjustment/zoom-fatfinger-expected.txt

Test that targetted area is picked even if the touch area is not fully contained.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS zoomableArea.top is 100
PASS zoomableArea.left is 100
PASS zoomableArea.width is 200
PASS zoomableArea.height is 20
PASS successfullyParsed is true

TEST COMPLETE
